code/app/requirements.txt
streamlit==1.54.0
requests==2.32.5
python-dotenv==1.2.1
opentelemetry-api==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-instrumentation-requests==0.60b1
opentelemetry-instrumentation-fastapi==0.60b1
opentelemetry-exporter-otlp-proto-http==1.39.1
